refactor intrinsic functions
