+++ tests/test_cf/test_unit/test_template.py
@@ -1,4 +1,3 @@
-import inspect
 from pathlib import Path
 from unittest.mock import mock_open, patch
 
@@ -7,10 +6,6 @@ import pytest
 from cloud_radar.cf.unit._template import (
     Template,
     add_metadata,
-    r_equals,
-    r_if,
-    r_ref,
-    r_sub,
 )
 
 
@@ -95,90 +90,6 @@ def test_render_false():
     assert "Foo" not in result["Resources"], "Resources should be empty."
 
     assert not result["Conditions"]["Bar"], "Condition should be false."
-
-
-def test_ref():
-
-    template = {"Parameters": {"foo": {"Value": "bar"}}}
-
-    add_metadata(template, Template.Region)
-
-    for i in inspect.getmembers(Template):
-        if not i[0].startswith("_"):
-            if not inspect.ismethod(i[1]):
-                result = r_ref(template, f"AWS::{i[0]}")
-                assert (
-                    result == i[1]
-                ), "Should be able to reference all pseudo variables."
-
-    result = r_ref(template, "foo")
-
-    assert result == "bar", "Should reference parameters."
-
-    result = r_ref(template, "SomeResource")
-
-    assert (
-        result == "SomeResource"
-    ), "If not a psedo var or parameter it should return the input."
-
-    fake = "AWS::FakeVar"
-
-    with pytest.raises(ValueError) as e:
-        r_ref(template, fake)
-
-    assert f"Unrecognized AWS Pseduo variable: '{fake}'." in str(e.value)
-
-
-def test_if():
-
-    template = {"Conditions": {"test": False}}
-
-    result = r_if(template, ["test", "true_value", "false_value"])
-
-    assert result == "false_value", "Should return the false value."
-
-    template["Conditions"]["test"] = True
-
-    result = r_if(template, ["test", "true_value", "false_value"])
-
-    assert result == "true_value", "Should return the true value."
-
-    with pytest.raises((Exception)):
-        # First value should the name of the condition to lookup
-        r_if(template, [True, "True", "False"])
-
-
-def test_equals():
-    # AWS is not very clear on what is valid here?
-    # > A value of any type that you want to compare.
-
-    true_lst = [True, "foo", 5, ["test"], {"foo": "foo"}]
-    false_lst = [False, "bar", 10, ["bar"], {"bar": "bar"}]
-
-    for idx, true in enumerate(true_lst):
-        false = false_lst[idx]
-
-        assert r_equals([true, true]), f"Should compare {type(true)} as True."
-
-        assert not r_equals([true, false]), f"Should compare {type(true)} as False."
-
-
-def test_sub():
-    template_dict = {"Parameters": {"Foo": {"Value": "bar"}}}
-
-    assert (
-        r_sub(template_dict, "Foo ${Foo}") == "Foo bar"
-    ), "Should subsuite a parameter."
-
-    result = r_sub(template_dict, "not ${!Test}")
-
-    assert result == "not ${Test}", "Should return a string literal."
-
-    add_metadata(template_dict, "us-east-1")
-
-    result = r_sub(template_dict, "${AWS::Region} ${Foo} ${!BASH_VAR}")
-
-    assert result == "us-east-1 bar ${BASH_VAR}", "Should render multiple variables."
 
 
 def test_resolve():
